backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.16,<3.17
django-cors-headers>=4.7,<4.8
django-ratelimit>=4.1,<4.2
scikit-learn>=1.6,<1.8
nltk>=3.9,<4.0
numpy>=2.2,<3.0
scipy>=1.15,<2.0
tldextract>=5.2,<5.4
